macro_te_scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
tabulate==0.9.0
PyYAML==6.0.1
python-dateutil==2.9.0.post0